# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ZVideo API
  version: 1.0.0
  description: REST API for a video hosting platform.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
...
  /videos:
    get:
      tags: [Videos]
      security: []
      summary: List videos
      operationId: listVideos
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Videos page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPage'
  /users/me/videos:
    get:
      tags: [Videos]
      summary: List current user's channel videos
      operationId: listMyVideos
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Videos page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /channels/{channelId}/videos:
    get:
      tags: [Videos]
      security: []
      summary: List channel videos
      operationId: listChannelVideos
      parameters:
        - $ref: '#/components/parameters/ChannelId'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Videos page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoPage'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/videos/uploads:
    post:
      tags: [Videos]
      summary: Initialize video upload
      operationId: initVideoUpload
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoUploadInitRequest'
      responses:
        '201':
          description: Pending video and upload URL created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoUploadInitResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{videoId}:
    get:
      tags: [Videos]
      security: []
      summary: Get video metadata
      operationId: getVideo
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: Video.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Videos]
      summary: Update video metadata
      operationId: updateVideo
      parameters:
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VideoUpdateRequest'
      responses:
        '200':
          description: Video updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Video'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Videos]
      summary: Delete video
      description: Channel owner, moderator, or admin can delete a video. File deletion may be performed asynchronously.
      operationId: deleteVideo
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '204':
          description: Video deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{videoId}/confirm:
    post:
      tags: [Videos]
      summary: Confirm uploaded video and publish it
      operationId: confirmVideoUpload
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '204':
          description: Video confirmed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{videoId}/stream-url:
    get:
      tags: [Videos]
      security: []
      summary: Get video streaming URL
      operationId: getVideoStreamUrl
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: Presigned URL for the original video file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PresignedUrlResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Video is not ready yet.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /videos/{videoId}/views:
    post:
      tags: [Videos]
      summary: Record a video view
      operationId: recordVideoView
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '204':
          description: View recorded.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{videoId}/rating:
    put:
      tags: [Videos]
      summary: Replace current user's video rating
      operationId: rateVideo
      parameters:
        - $ref: '#/components/parameters/VideoId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RatingRequest'
      responses:
        '204':
          description: Rating applied.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /videos/{videoId}/stats:
    get:
      tags: [Videos]
      security: []
      summary: Get video stats
      operationId: getVideoStats
      parameters:
        - $ref: '#/components/parameters/VideoId'
      responses:
        '200':
          description: Video stats.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VideoStats'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    VideoId:
      name: videoId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request validation or domain rule failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidUsername:
              summary: Invalid username
              value:
                error:
                  code: INVALID_USERNAME
                  message: The username is invalid
            invalidUserEmail:
              summary: Invalid email
              value:
                error:
                  code: INVALID_USER_EMAIL
                  message: The email is invalid
            weakPassword:
              summary: Weak password
              value:
                error:
                  code: WEAK_PASSWORD
                  message: The password is too weak
            playlistNameEmpty:
              summary: Empty playlist name
              value:
                error:
                  code: PLAYLIST_NAME_EMPTY
                  message: Playlist name cannot be empty
            playlistVideoChannelMismatch:
              summary: Playlist and video belong to different channels
              value:
                error:
                  code: PLAYLIST_VIDEO_CHANNEL_MISMATCH
                  message: Video belongs to another channel
            selfSubscription:
              summary: User cannot subscribe to own channel
              value:
                error:
                  code: SELF_SUBSCRIPTION
                  message: Cannot subscribe to your own channel
            invalidNotificationSettings:
              summary: Invalid notification settings
              value:
                error:
                  code: INVALID_NOTIFICATION_SETTINGS
                  message: Notification settings payload is invalid
            roleNotFound:
              summary: Role not found
              value:
                error:
                  code: ROLE_NOT_FOUND
                  message: Role not found
            communityPostContentEmpty:
              summary: Empty community post
              value:
                error:
                  code: COMMUNITY_POST_CONTENT_EMPTY
                  message: Community post content cannot be empty
            communityCommentContentEmpty:
              summary: Empty community comment
              value:
                error:
                  code: COMMUNITY_COMMENT_CONTENT_EMPTY
                  message: Community comment content cannot be empty
    Unauthorized:
      description: Authentication failed or token is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCredentials:
              summary: Invalid login credentials
              value:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
            invalidToken:
              summary: Invalid or expired token
              value:
                error:
                  code: INVALID_TOKEN
                  message: Invalid or expired token
            missingAuthorization:
              summary: Missing bearer token
              value:
                error:
                  code: UNAUTHORIZED
                  message: user not authorized
    Forbidden:
      description: Authenticated user is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: Insufficient permissions
              value:
                error:
                  code: FORBIDDEN
                  message: You don't have permission to access this resource
            userBanned:
              summary: User account is banned
              value:
                error:
                  code: USER_BANNED
                  message: User account is banned
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              summary: User not found
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User not found
            channelNotFound:
              summary: Channel not found
              value:
                error:
                  code: CHANNEL_NOT_FOUND
                  message: Channel not found
            videoNotFound:
              summary: Video not found
              value:
                error:
                  code: VIDEO_NOT_FOUND
                  message: Video not found
            playlistNotFound:
              summary: Playlist not found
              value:
                error:
                  code: PLAYLIST_NOT_FOUND
                  message: Playlist not found
            ratingNotFound:
              summary: Rating not found
              value:
                error:
                  code: RATING_NOT_FOUND
                  message: Rating not found
            commentNotFound:
              summary: Comment not found
              value:
                error:
                  code: COMMENT_NOT_FOUND
                  message: Comment not found
            communityPostNotFound:
              summary: Community post not found
              value:
                error:
                  code: COMMUNITY_POST_NOT_FOUND
                  message: Community post not found
            communityCommentNotFound:
              summary: Community comment not found
              value:
                error:
                  code: COMMUNITY_COMMENT_NOT_FOUND
                  message: Community comment not found
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Human-readable error message
...
    VideoStatus:
      type: string
      enum: [pending, ready]
    Video:
      type: object
      required: [id, channelId, channelName, title, description, filepath, originalFilename, status, createdAt]
      properties:
        id:
          type: integer
        channelId:
          type: integer
        channelName:
          type: string
        title:
          type: string
        description:
          type: string
        filepath:
          type: string
          description: Storage object key.
        originalFilename:
          type: string
        status:
          $ref: '#/components/schemas/VideoStatus'
        createdAt:
          type: string
          format: date-time
        stats:
          $ref: '#/components/schemas/VideoStats'
    VideoStats:
      type: object
      required: [views, likes, dislikes, comments]
      properties:
        views:
          type: integer
          minimum: 0
        likes:
          type: integer
          minimum: 0
        dislikes:
          type: integer
          minimum: 0
        comments:
          type: integer
          minimum: 0
    VideoUploadInitRequest:
      type: object
      required: [title, filename]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
          default: ''
        filename:
          type: string
          minLength: 1
          example: video.mp4
    VideoUploadInitResponse:
      type: object
      required: [video, uploadUrl, uploadExpiresAt]
      properties:
        video:
          $ref: '#/components/schemas/Video'
        uploadUrl:
          type: string
          format: uri
        uploadExpiresAt:
          type: string
          format: date-time
    VideoUpdateRequest:
      type: object
      minProperties: 1
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 64
        description:
          type: string
    PresignedUrlResponse:
      type: object
      required: [url, expiresAt]
      properties:
        url:
          type: string
          format: uri
        expiresAt:
          type: string
          format: date-time
    RatingAction:
      type: string
      enum: [like, dislike, remove]
    RatingRequest:
      type: object
      required: [action]
      properties:
        action:
          $ref: '#/components/schemas/RatingAction'
...
    PageMeta:
      type: object
      required: [limit, offset, totalCount]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        totalCount:
          type: integer
          format: int64
          minimum: 0
...
    VideoPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Video'
...